backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.29
psycopg2-binary==2.9.9

python-dotenv==1.0.1

passlib[bcrypt]==1.7.4
python-jose==3.3.0

pydantic==2.6.4
pydantic-settings==2.2.1

email-validator==2.1.1
